workShopCode/workshopMaterials/workshop/labs/step5/nonTrivialProject/requirements.txt
coverage==4.5.3
Django==2.1.7
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
